+++ .storybook/theme-addon/register.tsx
@@ -1,0 +1,26 @@
+import * as React from 'react';
+import {addons, types} from '@storybook/addons';
+import {useGlobals, type API} from '@storybook/api';
+import {getThemeType} from '@gravity-ui/uikit';
+import {themes} from '../theme';
+
+const ADDON_ID = 'yc-theme-addon';
+const TOOL_ID = `${ADDON_ID}tool`;
+
+addons.register(ADDON_ID, (api) => {
+    addons.add(TOOL_ID, {
+        type: types.TOOL,
+        title: 'Theme',
+        render: () => {
+            return <Tool api={api} />;
+        },
+    });
+});
+
+function Tool({api}: {api: API}) {
+    const [{theme}] = useGlobals();
+    React.useEffect(() => {
+        api.setOptions({theme: themes[getThemeType(theme)]});
+    }, [theme]);
+    return null;
+}