+++ src/components/AsideHeader/AsideHeader.tsx
@@ -11,11 +11,17 @@ import {Drawer, DrawerItem, DrawerItemProps} from '../Drawer/Drawer';
 import {Logo} from '../Logo/Logo';
 import {CompositeBar} from '../CompositeBar/CompositeBar';
 import {Content, RenderContentType} from '../Content';
+import {fakeDisplayName} from '../helpers';
 
 import controlMenuButtonIcon from '../../../assets/icons/control-menu-button.svg';
 import headerDividerCollapsedIcon from '../../../assets/icons/divider-collapsed.svg';
 
+import {AsideHeaderContextProvider} from './AsideHeaderContext';
+
 import './AsideHeader.scss';
+
+// TODO: remove temporary fix for expand button
+const NotIcon = fakeDisplayName('NotIcon', Icon);
 
 const b = block('aside-header');
 
@@ -64,18 +70,19 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
         const size = compact ? ASIDE_HEADER_COMPACT_WIDTH : ASIDE_HEADER_EXPANDED_WIDTH;
 
         return (
-            <div className={b({compact}, className)}>
-                <div className={b('pane-container')}>
-                    {this.renderFirstPane(size)}
-                    {this.renderSecondPane(size)}
+            <AsideHeaderContextProvider value={{compact, size}}>
+                <div className={b({compact}, className)}>
+                    <div className={b('pane-container')}>
+                        {this.renderFirstPane(size)}
+                        {this.renderSecondPane(size)}
+                    </div>
                 </div>
-            </div>
+            </AsideHeaderContextProvider>
         );
     }
 
     private renderFirstPane = (size: number) => {
-        const {dict, menuItems, panelItems, compact, headerDecoration, multipleTooltip} =
-            this.props;
+        const {dict, menuItems, panelItems, headerDecoration, multipleTooltip} = this.props;
 
         return (
             <React.Fragment>
@@ -86,7 +93,6 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
                         {menuItems?.length ? (
                             <CompositeBar
                                 items={menuItems}
-                                compact={compact}
                                 enableCollapsing={true}
                                 dict={dict}
                                 onItemClick={this.onItemClick}
@@ -115,9 +121,7 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
         );
     };
 
-    private renderLogo = () => (
-        <Logo {...this.props.logo} compact={this.props.compact} onClick={this.onLogoClick} />
-    );
+    private renderLogo = () => <Logo {...this.props.logo} onClick={this.onLogoClick} />;
 
     private renderHeader = () => (
         <div className={b('header', {['with-decoration']: this.props.headerDecoration})}>
@@ -125,7 +129,6 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
 
             <CompositeBar
                 items={this.props.subheaderItems}
-                compact={this.props.compact}
                 enableCollapsing={false}
                 onItemClick={this.onItemClick}
             />
@@ -140,7 +143,7 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
     );
 
     private renderFooter = (size: number) => {
-        const {compact, renderFooter} = this.props;
+        const {renderFooter, compact} = this.props;
 
         return (
             <div className={b('footer')}>
@@ -171,7 +174,7 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
     };
 
     private renderCollapseButton = () => {
-        const {compact, dict} = this.props;
+        const {dict, compact} = this.props;
         const typeButton = compact ? Dict.ExpandButton : Dict.CollapseButton;
 
         return (
@@ -181,7 +184,7 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
                 onClick={this.onCollapseButtonClick}
                 title={dict?.[typeButton] ?? defaultDict[typeButton]}
             >
-                <Icon
+                <NotIcon
                     data={controlMenuButtonIcon}
                     className={b('collapse-icon')}
                     width="16"
@@ -199,9 +202,13 @@ export class AsideHeader extends React.Component<AsideHeaderInnerProps> {
         this.props.onClosePanel?.();
     };
 
-    private onItemClick = (item: MenuItem, collapsed: boolean) => {
+    private onItemClick = (
+        item: MenuItem,
+        collapsed: boolean,
+        event: React.MouseEvent<HTMLDivElement, MouseEvent>,
+    ) => {
         this.props.onClosePanel?.();
-        item.onItemClick?.(item, collapsed);
+        item.onItemClick?.(item, collapsed, event);
     };
 
     private onLogoClick = (event: React.MouseEvent<HTMLElement, MouseEvent>) => {