+++ src/components/CompositeBar/__stories__/CompositeBar.stories.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
-import {Meta} from '@storybook/react/types-6-0';
-import {Story} from '@storybook/react';
+
+import type {Meta, StoryFn} from '@storybook/react';
 
 import {CompositeBar, CompositeBarProps} from '../CompositeBar';
 import {ASIDE_HEADER_COMPACT_WIDTH, ASIDE_HEADER_EXPANDED_WIDTH} from '../../constants';
 import {menuItemsShowcase} from './moc';
+
+import {AsideHeaderContextProvider} from '../../AsideHeader/AsideHeaderContext';
 
 import './CompositeBarShowcase.scss';
 
@@ -27,14 +29,15 @@ export default {
     ],
 } as Meta;
 
-const Template: Story<CompositeBarProps> = (args) => (
+const Template: StoryFn<CompositeBarProps> = (args) => (
     <div className="composite-bar-showcase">
-        <CompositeBar {...args} />
+        <AsideHeaderContextProvider value={{compact: false, size: ASIDE_HEADER_EXPANDED_WIDTH}}>
+            <CompositeBar {...args} />
+        </AsideHeaderContextProvider>
     </div>
 );
 
 export const Default = Template.bind({});
 Default.args = {
-    compact: false,
     items: menuItemsShowcase,
 };