+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export {AsideHeader, AsideHeaderProps} from './AsideHeader/AsideHeader';
+export {AsideHeaderContextProvider, useAsideHeaderContext} from './AsideHeader/AsideHeaderContext';
 export {Drawer, DrawerProps, DrawerItemProps, DrawerItem} from './Drawer/Drawer';
 export {FooterItem, FooterItemProps} from './FooterItem/FooterItem';
 export * from './ActionBar';
@@ -7,3 +8,5 @@ export * from './HotkeysPanel';
 export * from './Settings';
 export * from './MobileHeader';
 export * from './types';
+
+export {Lang, configure} from './utils/configure';