+++ src/components/types.ts
@@ -17,7 +17,12 @@ export interface MenuItem {
     link?: string;
     current?: boolean;
     pinned?: boolean;
-    onItemClick?: (item: MenuItem, collapsed: boolean) => void;
+    onItemClick?: (
+        item: MenuItem,
+        collapsed: boolean,
+        event: React.MouseEvent<HTMLDivElement, MouseEvent>,
+    ) => void;
+    onItemClickCapture?: (event: React.SyntheticEvent) => void;
     itemWrapper?: (
         p: MakeItemParams,
         makeItem: (p: MakeItemParams) => React.ReactNode,